+++ CHANGELOG.md
@@ -6,6 +6,8 @@ the structure of [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## [Unreleased]
 
+## [0.20.8] - 2026-08-25
+
 ### Fixed
 
 - Restored distinct populated lifecycle states in **Preview All** and **Send Test